+++ pyproject.toml
@@ -34,4 +34,4 @@ Issues   = "https://github.com/your-handle/spectr/issues"
 Docs     = "https://spectr.readthedocs.io"
 
 [project.scripts]
-spectr = "spectr.spectr:main"   # $ spectr ↵ launches the UI
+spectr = "spectr.cli:main"   # $ spectr ↵ launches the UI
